Primeiro commit

+++ app/controllers/api/v1/accounts/conversations/messages_controller.rb
@@ -9,8 +9,16 @@ class Api::V1::Accounts::Conversations::MessagesController < Api::V1::Accounts::
     user = Current.user || @resource
     mb = Messages::MessageBuilder.new(user, @conversation, params)
     @message = mb.perform
+    update_conversation_status
   rescue StandardError => e
     render_could_not_create_error(e.message)
+  end
+
+  def update_conversation_status
+    # change conversation status to open if resolved or snoozed
+    return unless @conversation.status == 'resolved' || @conversation.status == 'snoozed'
+
+    @conversation.update!(status: :open)
   end
 
   def update
@@ -19,6 +27,8 @@ class Api::V1::Accounts::Conversations::MessagesController < Api::V1::Accounts::
   end
 
   def destroy
+    return head :bad_request unless message.can_delete_message?
+
     ActiveRecord::Base.transaction do
       message.update!(content: I18n.t('conversations.messages.deleted'), content_type: :text, content_attributes: { deleted: true })
       message.attachments.destroy_all
@@ -54,6 +64,13 @@ class Api::V1::Accounts::Conversations::MessagesController < Api::V1::Accounts::
     render json: { content: translated_content }
   end
 
+  def forward
+    ::Conversations::ForwardMessageJob.perform_later(forward_message_params)
+    head :ok
+  rescue StandardError => e
+    render e
+  end
+
   private
 
   def message
@@ -70,6 +87,15 @@ class Api::V1::Accounts::Conversations::MessagesController < Api::V1::Accounts::
 
   def already_translated_content_available?
     message.translations.present? && message.translations[permitted_params[:target_language]].present?
+  end
+
+  def forward_message_params
+    {
+      user_id: Current.user.id,
+      account_id: Current.account.id,
+      message_id: message.id,
+      contacts: params[:contacts]
+    }
   end
 
   # API inbox check
